* master: (752 commits)
  fix default translation - closes #13034
  fix puzzle board menu on mobile - closes #13028
  remove coach review translations
  log oauth prompts
  make the mobile oauth authorization page less scary
  fix oauth with signed token
  more controller code golf
  controller code golf
  delete json export
  no credentialless on Plan.index
  live setting for credentialless UA matcher
  try disabling credentialless
  New Crowdin updates (#13022)
  +Artsakh Republic flag #11631
  ~Fix daily puzzle ui jumping when toggling #12960
  given Conversion[scalatags.Text.Frag, Fu[Result]]
  export reviews as json
  refactor LilaController
  support show moves as a voice command
  log unsigned oauth bearers
  ...

+++ modules/i18n/src/main/package.scala
@@ -4,7 +4,7 @@ import play.api.i18n.Lang
 
 export lila.Lila.{ *, given }
 
-private type Count      = Int
+private type Count      = Long
 private type MessageKey = String
 private type MessageMap = java.util.Map[MessageKey, Translation]
 private type Messages   = Map[Lang, MessageMap]
